Moved python binding util code from tests to bindings
